Feat/timeouts

+++ tavily/errors.py
@@ -17,7 +17,7 @@ class InvalidAPIKeyError(Exception):
         super().__init__(message)
 
 class TimeoutError(Exception):
-    def __init__(self, timeout: int):
+    def __init__(self, timeout: float):
         super().__init__(f"Request timed out after {timeout} seconds.")
 
 class MissingAPIKeyError(Exception):
